refactor(gui): isolate backend project model
